{{cookiecutter.project_slug}}/requirements/base.txt
Django==5.1
django-environ==0.8.1
django-model-utils==4.2.0

djangorestframework==3.15.2
django-filter==21.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.11.0

{%- if cookiecutter.use_whitenoise == "y" %}
whitenoise==6.0.0
{%- endif %}

redis==4.1.4

{%- if cookiecutter.use_dramatiq == "y" %}
dramatiq==1.17.0
git+https://github.com/cuongnb14/django_dramatiq@v1.1.3#egg=django_dramatiq
{%- endif %}

ipython
gunicorn==20.1.0
honcho==1.1.0

# linter
flake8-colors
flake8

drf-spectacular==0.22.1

git+https://github.com/cuongnb14/django-admin-extended.git@v5.1.1#egg=django-admin-extended

pytest-django==4.9.0
pytest-xdist==3.6.1
requests==2.32.3
faker==26.0.0
tzdata==2024.1

mixer==7.2.2
